requirements.txt
Flask==3.1.1
flask-wtf==1.2.2
google-adk==1.7.0
google-generativeai==0.8.5
google-cloud-aiplatform==1.104.0
marshmallow==3.26.1
python-dotenv==1.1.1
pytest==8.4.1
pytest-cov==6.2.1
flask-jwt-extended==4.6.0
flask-limiter==3.5.0
flask-talisman==1.1.0
pyOpenSSL==24.0.0
gunicorn==21.2.0
pydantic>=2.6.0,<2.10.0
safety==3.6.0
python-dotenv==1.1.1
llama-index>=0.10.0